requirements.txt
py7zr
dictor
pysftp
pyyaml
fs
python-gnupg
jinja2